+++ modules/engines/engine-hdb/src/main/java/com/sap/xsk/hdb/ds/processors/table/XSKTableCreateProcessor.java
@@ -15,8 +15,8 @@ import com.google.inject.Inject;
 import com.sap.xsk.hdb.ds.api.IXSKDataStructureModel;
 import com.sap.xsk.hdb.ds.model.hdbtable.XSKDataStructureHDBTableModel;
 import com.sap.xsk.hdb.ds.processors.AbstractXSKProcessor;
+import com.sap.xsk.hdb.ds.processors.table.utils.XSKTableEscapeService;
 import com.sap.xsk.hdb.ds.service.manager.IXSKDataStructureManager;
-import com.sap.xsk.hdb.ds.processors.table.utils.XSKTableCreateEscapeService;
 import com.sap.xsk.utils.XSKConstants;
 import com.sap.xsk.utils.XSKHDBUtils;
 import org.eclipse.dirigible.database.sql.DatabaseArtifactTypes;
