+++ modules/engines/engine-hdb/src/main/java/com/sap/xsk/hdb/ds/processors/table/utils/XSKTableAlterHandler.java
@@ -1,12 +1,12 @@
 /*
- * Copyright (c) 2019-2021 SAP SE or an SAP affiliate company and XSK contributors
+ * Copyright (c) 2021 SAP SE or an SAP affiliate company and XSK contributors
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Apache License, v2.0
  * which accompanies this distribution, and is available at
  * http://www.apache.org/licenses/LICENSE-2.0
  *
- * SPDX-FileCopyrightText: 2019-2021 SAP SE or an SAP affiliate company and XSK contributors
+ * SPDX-FileCopyrightText: 2021 SAP SE or an SAP affiliate company and XSK contributors
  * SPDX-License-Identifier: Apache-2.0
  */
 package com.sap.xsk.hdb.ds.processors.table.utils;
@@ -184,7 +184,7 @@ public class XSKTableAlterHandler {
       dropExistingIndex(connection, stmt, droppedIndices, rsIndeces);
     }
 
-    XSKTableCreateEscapeService escapeService = new XSKTableCreateEscapeService(connection, this.tableModel);
+    XSKTableEscapeService escapeService = new XSKTableEscapeService(connection, this.tableModel);
 
     escapeService.escapeTableBuilderUniqueIndices(alterTableBuilder);
     executeAlterBuilder(connection, alterTableBuilder);
